Implement Product class
