+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Offer.java
@@ -38,16 +38,17 @@ public class Offer {
         if (obj == null) {
             return false;
         }
-		
+
         if (getClass() != obj.getClass()) {
             return false;
         }
-		
+
         Offer other = (Offer) obj;
         if (availableItems == null) {
-            if (other.availableItems != null) 
+            if (other.availableItems != null) {
                 return false;
-            
+            }
+
         } else if (!availableItems.equals(other.availableItems)) {
             return false;
         }
@@ -67,7 +68,7 @@ public class Offer {
         }
 
         for (OfferItem item : availableItems) {
-            OfferItem sameItem = seenOffer.findItem(item.getProductId());
+            OfferItem sameItem = seenOffer.findItem(item.getProduct().getProductId());
             if (sameItem == null) {
                 return false;
             }
@@ -81,7 +82,7 @@ public class Offer {
 
     private OfferItem findItem(String productId) {
         for (OfferItem item : availableItems) {
-            if (item.getProductId().equals(productId)) {
+            if (item.getProduct().getProductId().equals(productId)) {
                 return item;
             }
         }
